More app logic merge conflicts that had to be fixed
